Add buttons to creation form
This is likely to change. The core files seem inconsistent and there's
confusion as far as how to call the controller (IOW there's more than
one way to accomplish this). For now, blocks are
going to be leveraged to create paths for controllers. There are quite a
few unkowns to this approach.