Add Cron actions datasource
The <actionsColumn> node in the ui component xml file depends on this
class. It provdes the action urls to the "Action" column by iterating
through each row in the listing UI component and adding the approprate
data.

It's important to note that when creating UI components, it is done
recursively via Magento\Framework\View\Element\UiComponent\Context::prepareDataSource()
NOT TO BE CONFUSED WITH the prepareDataSource() method implemented in
the Magento\Framework\View\Element\UiComponentInterface.

Lastly, it should be noted that Context::prepareDataSource() is the one
doing the recursive method calling to all objects that implement the
UiComponentInterface. This is a very important distinction between the
two classes.