Fix timezone convertion bug
This was quite troublesome becuase magento calculates all date values from
the server as UTC and then calculates the timezone offset from the
store timezone (this calculation is done by moment.js). This isn't an issue for
values that are saved as UTC (as almost all date values are in magento), but magento does NOT save cron jobs as
UTC; they're saved (based on my findings in the core files) based on
the store timezone.

So by telling magento the storeTimezone is 'UTC' in the Javascript
configurations (aka uiCompoenent XML), it will not calculate the offset
in the edit page, causing moment.js to not calculate the offset.