Create base for repurposed timeline component
After digging through the magento Ui module, I was able to locate a
timeline component. I decided that rather than starting from scratch,
this component can be repurposed. Albeit, this has proven to be very
difficult to work with, specifically how time is distributed
proportionally through the component.

There are a lot of draw backs to this decision, but I believe this could
prove to be more powerful than other timelines because of the powerful
filtering that can be leveraged.

Currently, the timeline is functional, but the data is completely
incorrect. There needs to be additional stylizations, so I created a
less module for future styling.