Create custom timeline component
This will allow us to create a custom uiComponent and data bind the
result of the ajax call (returning json) to the imported variable
"rows". The data will be indexed by the name of the job code (sorted).

We should be able to iterate through this and instantiate rows and
data points as needed