Add proxy to avoid loading StoreManager
During the setup:install process the list of CLI commands is called.
One of these commands has a dependency on the StoreManagerInterface.
This tries to fetch database from the database from its constructor
which causes an error during integration tests because the table hasn't
been created yet. Adding a proxy delays this call to the DB until it is
used.