Remove unused code
The function this was wrapping was introduced in v2.2.0-RC1.2 and was called
within Magento\Cron\Model\Schedule::tryLockJob(), which we also have an after
plugin for. As we are setting (and now saving, see #142) pid & hostname in
there, this around plugin is no longer required. Since v2.3.5, tryLockJob() no
longer calls this function, further rendering this plugin obsolete.